17.04.2023

+++ jaut.proj/index2.py
@@ -4,22 +4,20 @@ from tkinter import filedialog
 from tkinter import ttk
 import sqlite3
 
+edit = False
+edit_id = 0
+
 # Create table
 conn = sqlite3.connect("datasheet.db")
 cursor = conn.cursor()
 cursor.execute("CREATE TABLE IF NOT EXISTS datasheet (id INTEGER PRIMARY KEY, type VARCHAR, model VARCHAR, price FLOAT)")
 conn.commit()
-conn.close()
 
 # Display data in Tree widget
 def display():
     tree.delete(*tree.get_children())
-    conn = sqlite3.connect("datasheet.db")
-    cursor = conn.cursor()
     cursor.execute("SELECT * FROM datasheet")
     parts = cursor.fetchall()
-    conn.commit()
-    conn.close()
     
     for part in parts:
         tree.insert('', 'end', text="2", values=(part[0], part[1], part[2], part[3]))
@@ -27,29 +25,74 @@ def display():
 
 # Submit
 def submit():
-    conn = sqlite3.connect("datasheet.db")
     cursor = conn.cursor()
     if type.get() != "" and name.get() != "" and price.get() != "":
         cursor.execute("INSERT INTO datasheet (type, model, price) VALUES (?, ?, ?)", (type.get(), name.get(), float(price.get())))
+    else:
+        messagebox.showinfo("error", "tukša ievade")
     conn.commit()
-    conn.close()
+    type.insert(0, "")
+    name.insert(0, "")
+    price.insert(0, "")
     display()
 
 # Delete item
 def delete():
     treeIndex = tree.focus()
-    id = tree.item(treeIndex)["values"][0]
-    cursor.execute("DELETE FROM datasheet WHERE id = ?",str(id))
-    conn.commit()
-    display()
+    if treeIndex:
+        id = tree.item(treeIndex)["values"][0] # focus atgriež indeksu, kuru izmanto, lai dabūtu konkrēto item
+        cursor.execute("DELETE FROM datasheet WHERE id = ?", str(id))
+        conn.commit()
+        display()
+    else:
+        messagebox.showinfo("Error","atzīmē dzēšamo vērtību")
 
-    # print(tree.item()) # focus atgriež indeksu, kuru izmanto, lai dabūtu konkrēto item
+def update():
+    global edit
+    global edit_id
+    
+    if not edit:
+        treeIndex = tree.focus()
+        if treeIndex:
+            edit_id = tree.item(treeIndex)["values"][0] # focus atgriež indeksu, kuru izmanto, lai dabūtu konkrēto item
+            cursor.execute("SELECT * FROM datasheet WHERE id = ?", str(edit_id))
+            part = cursor.fetchone()
+            
+            type.insert(0, part[1])
+            name.insert(0, part[2])
+            price.insert(0, part[3])
+            edit = True
+        else:
+            messagebox.showinfo('error',"tu i tups")
+    else:
+        if type.get() != "" and name.get() != "" and price.get() != "":
+            cursor.execute("UPDATE datasheet SET type = ?, model = ?, price = ? WHERE id = ?", (type.get(), name.get(), float(price.get()), str(edit_id)))
+            conn.commit()
+            type.delete(0, len(type.get()))
+            name.delete(0, len(name.get()))
+            price.delete(0, len(price.get()))
+            display()
+            edit = False
 
+# print out items
+def printout():
+    treeIndex = tree.focus()
+    if treeIndex:
+        edit_id = tree.item(treeIndex)["values"][0] # focus atgriež indeksu, kuru izmanto, lai dabūtu konkrēto item
+        cursor.execute("SELECT * FROM datasheet WHERE id = "+ str(edit_id))
+        part = cursor.fetchone()
+        with open("printout"+ str(part[0]),"w") as f:
+            f.write("personala datora sastavdala \n")
+            f.write('veids: ' + part[1] )
+            f.write("\nmodelis: "+ part[2])
+            f.write("\ncena: " + str(part[3]))
+    else:
+        messagebox.showinfo("error","ludzu, atzīmē vēlamo printējamo")
 
 # Window configuration
 window = Tk()
 
-window.geometry("800x400")
+window.geometry("800x500")
 window.title("Datora sastavdalas")
 window.config(background="white")
 
@@ -94,15 +137,14 @@ submitb = Button(window,
 
 submitb.place(x=330, y= 140)
 
-# Save List Button (Need def)
-# save = Button(window,
-#               font=("Visby",13, "bold"),
-#               text="save list",
-#               background="Green",
-#               activebackground="Green",
-#               command=listsave)
+printb = Button(window,
+                font=("Visby",13, "bold"),
+                text="print",
+                background="#FFFFFF",
+                activebackground="#94a0a8",
+                command=printout)
 
-# save.place(x=410, y= 72)
+printb.place(x=650, y=140)
 
 tree = ttk.Treeview(window, column=("c1", "c2", "c3", "c4"), show='headings', height=8)
 
@@ -127,16 +169,18 @@ ldelete = Button(window,
 
 ldelete.place(x=450,y=140)
 
-ldelete = Button(window,
+edit = Button(window,
                  font=("Visby",13, "bold"),
                  text="edit",
                  background="White",
                  activebackground="White",
-                 command=delete)
+                 command=update)
 
-ldelete.place(x=550,y=140)
-
+edit.place(x=550,y=140)
 
 # Program executer
+edit = False
+edit_id = 0
 display()
 window.mainloop()
+conn.close()
